refactor: 남은 시간 리팩토링
