+++ constants/time.ts
@@ -1,0 +1,1 @@
+export const OTP_TIME = 3600;
